Fixing console warnings by adding key attribute
Adding key attribute is essential when we are dealing with lists in
React. This will helps React to handle the lists updates. When the key
is not supplied, React needs to update/reload the whole list every time
in order to present the data properly and updated in the page. Besides
causing some errors, it also may cause performance issues. So, use the
key always when you have to deal with lists in React.